requirements.txt
Django>=5.0,<6.0
python-dotenv>=1.0
pydantic>=2.8
chroma-hnswlib>=0.7
chromadb>=0.5
sentence-transformers>=3.0
google-genai>=0.3.0
pymupdf>=1.24.9
python-docx>=1.1.2
pillow>=10.4.0
pytesseract>=0.3.13
uvicorn>=0.30
gunicorn>=22.0
whitenoise>=6.7
requests>=2.31
